+++ app/views/support_interface/application_forms/show.html.erb
@@ -32,3 +32,13 @@
     </dd>
   </div>
 </dl>
+
+<% if @application_form.submitted? %>
+  <h2 class="govuk-heading-l govuk-!-margin-top-8" id="first-reference">First reference</h2>
+
+  <%= render ReferenceWithFeedbackComponent, reference: @application_form.references.first %>
+
+  <h2 class="govuk-heading-l govuk-!-margin-top-8" id="second-reference">Second reference</h2>
+
+  <%= render ReferenceWithFeedbackComponent, reference: @application_form.references.second %>
+<% end %>
